expose notification choice to consumers
