+++ src/Altinn.App.Core/Features/Signing/Models/ContentWrapper.cs
@@ -1,10 +1,17 @@
 using Altinn.App.Core.Features.Correspondence.Models;
+using Altinn.App.Core.Features.Signing.Enums;
+using Altinn.App.Core.Models;
 
 namespace Altinn.App.Core.Features.Signing.Models;
 
 internal record ContentWrapper
 {
     internal required CorrespondenceContent CorrespondenceContent { get; init; }
+    internal required string SendersReference { get; init; }
+    internal OrganisationOrPersonIdentifier? Recipient { get; init; }
+    internal NotificationChoice? NotificationChoice { get; init; }
+    internal Notification? Notification { get; init; }
+    internal Notification? ReminderNotification { get; init; }
     internal string? SmsBody { get; init; }
     internal string? EmailBody { get; init; }
     internal string? EmailSubject { get; init; }
